examples: add support for joining and leaving public groups, groundwork for owners adding/removing for private groups

+++ examples/backend-cloudflare/src/routes/xrpc/groups.ts
@@ -24,12 +24,14 @@ const GROUP_COLLECTION = 'app.blebbit.authr.group'
 export function addRoutes(app: Hono) {
   app.get('/xrpc/app.blebbit.authr.getGroup', getGroup)
   app.get('/xrpc/app.blebbit.authr.getGroups', getGroups)
+
   app.post('/xrpc/app.blebbit.authr.createGroup', createGroup)
+  // app.post('/xrpc/app.blebbit.authr.updateGroup', updateGroup)
   app.post('/xrpc/app.blebbit.authr.deleteGroup', deleteGroup)
 
   app.post('/xrpc/app.blebbit.authr.addGroupMember', addGroupMember)
   app.post('/xrpc/app.blebbit.authr.setGroupMember', setGroupMember)
-  app.post('/xrpc/app.blebbit.authr.rmvGroupMember', rmvGroupMember)
+  app.post('/xrpc/app.blebbit.authr.removeGroupMember', removeGroupMember)
 }
 
 async function getGroup(c: Context) {
@@ -172,47 +174,90 @@ async function createGroup(c: Context) {
 }
 
 
-async function checkIfOwner(c: Context) {
+async function canWriteMember(c: Context, group: any, role: string, newDid: string, reqDid: string) {
+  var canMod = false
+  // anyone can add themselves to a public group
+  if (group.public && newDid === reqDid && role === 'member') {
+    canMod = true
+  }
+
+  if (!canMod) {
+    // check if the requester is an owner of the group
+    const check = await checkPermission(c.env, `blog/group:${group.id}`, "owner", `blog/user:${reqDid.replaceAll(":", "_")}`)
+
+    console.log("canWriteMember.ownerCheck", check)
+
+    if (check && check.response && check.response.item?.permissionship === 2) {
+      canMod = true
+    }
+  }
+
+  return canMod
+}
+
+async function getGroupFromDB(c: Context, groupId: string) {
+  const groupDb =
+    await c.env.DB
+    .prepare('SELECT * FROM records WHERE nsid = ? AND id = ?')
+    .bind(GROUP_COLLECTION, groupId)
+    .all()
+
+  var groups = groupDb.results as any[]
+  console.log("addGroupMember.groups", groups)
+  if (!groups || groups.length === 0) {
+    return null
+  }
+
+  return groups[0]
+}
+
+async function commonGroupMemberChecks(c: Context) {
   const authrSession = c.get("authrSession")
   const pdsSession = c.get("pdsSession")
-
-  const payload = await c.req.json()
-  console.log("createGroup.payload", payload)
-  const { groupId, role, did: newDid } = payload
-
-  // who's writing this group?
-  var did =  pdsSession?.iss || authrSession?.did || undefined
-  console.log("createGroup.did", did)
+  const reqDid =  pdsSession?.iss || authrSession?.did || undefined
 
   // must be authenticated to perform writes of any kind
-  if (!did) {
+  if (!reqDid) {
     return c.json({
       error: 'Not authenticated',
     }, 401)
   }
 
-  // must be a group owner to add members
-  const check = await checkPermission(c.env, `blog/group:${groupId}`, "owner", `blog/user:${did.replaceAll(":", "_")}`)
+  // unpack the payload
+  const payload = await c.req.json()
+  console.log("addGroupMember.payload", payload)
+  const { groupId, role, did: newDid } = payload
 
-  console.log("addGroupMember.check", check)
-
-  if (!check || !check.response || check.response.item?.permissionship !== 2) {
+  // lookup the group in the database
+  const group = await getGroupFromDB(c, groupId)
+  if (!group) {
     return c.json({
-      error: 'Not authorized',
-    }, 403)
+      error: 'Group not found',
+    }, 404)
   }
 
-  return payload
-
+  const canAdd = await canWriteMember(c, group, role, newDid, reqDid)
+  if (!canAdd) {
+    return c.json({
+      error: 'Not authorized to add members to this group',
+    }, 403)
+  }
+  return {
+    groupId,
+    role,
+    newDid,
+    reqDid,
+  }
 }
 
 async function addGroupMember(c: Context) {
-  const result = await checkIfOwner(c)
+ const result = await commonGroupMemberChecks(c) as any
   if (result instanceof Response) {
-    return result // if we got a response, return it
+    return result
   }
+  console.log("addGroupMember.payload", result)
+  const { groupId, role, newDid } = result
 
-  const { groupId, role, did: newDid } = result
   // add the member to the group
   const perm = await createRelationship(c.env, `blog/group:${groupId}`, role, `blog/user:${newDid.replaceAll(":", "_")}`)
   console.log("addGroupMember.perm", perm)
@@ -221,27 +266,36 @@ async function addGroupMember(c: Context) {
 }
 
 async function setGroupMember(c: Context) {
-  const result = await checkIfOwner(c)
+ const result = await commonGroupMemberChecks(c) as any
   if (result instanceof Response) {
-    return result // if we got a response, return it
+    return result
   }
+  console.log("setGroupMember.payload", result)
+  const { groupId, role, newDid } = result
 
-  const { groupId, role, did: newDid } = result
-  // add the member to the group
+  // update member of the group
   const perm = await updateRelationship(c.env, `blog/group:${groupId}`, role, `blog/user:${newDid.replaceAll(":", "_")}`)
   console.log("addGroupMember.perm", perm)
 
   return c.json(perm.response, 200)
 }
 
-async function rmvGroupMember(c: Context) {
-  const result = await checkIfOwner(c)
+async function removeGroupMember(c: Context) {
+  const result = await commonGroupMemberChecks(c) as any
   if (result instanceof Response) {
-    return result // if we got a response, return it
+    return result
+  }
+  console.log("addGroupMember.payload", result)
+  const { groupId, role, newDid, reqDid } = result
+
+  // todo, owner should not be able to remote thmeselves
+  if (role === 'owner' && reqDid === newDid) {
+    return c.json({
+      error: 'Cannot remove yourself as owner',
+    }, 403)
   }
 
-  const { groupId, role, did: newDid } = result
-  // add the member to the group
+  // remove member from the group
   const perm = await deleteRelationship(c.env, `blog/group:${groupId}`, role, `blog/user:${newDid.replaceAll(":", "_")}`)
   console.log("addGroupMember.perm", perm)
 
@@ -249,12 +303,37 @@ async function rmvGroupMember(c: Context) {
 }
 
 async function deleteGroup(c: Context) {
-  const result = await checkIfOwner(c)
-  if (result instanceof Response) {
-    return result // if we got a response, return it
+
+  const check = await commonGroupMemberChecks(c) as any
+  if (check instanceof Response) {
+    return check
+  }
+  console.log("addGroupMember.payload", check)
+  const { groupId, reqIsOwner } = check
+
+  // todo, owner should not be able to remote thmeselves
+  if (!reqIsOwner) {
+    return c.json({
+      error: 'Cannot delete group, you are not an owner',
+    }, 403)
   }
 
+  // first remove from database
+  const result =
+    await c.env.DB
+    .prepare('SELECT * FROM records WHERE nsid = ? AND id = ?') 
+    .bind(GROUP_COLLECTION, groupId)
+    .all()
+  var groups = result.results as any[]
+  console.log("getGroups.groups", groups)
+
+
+  // then remove permissions tied to the group
+  const perm = await deleteRelationship(c.env, `blog/group:${groupId}`, undefined, undefined)
+
+  c.status(204)
   return c.json({
-    error: 'Not implemented',
-  }, 501)
+    message: 'Group deleted successfully',
+    perm,
+  })
 }
